refactor(app): reorganize existing files into proper app, core, providers, and screens structure
