+++ lib/app/app.dart
@@ -1,0 +1,56 @@
+// lib/app/app.dart
+import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'app_theme.dart';
+import 'app_routes.dart';
+import '../providers/theme_provider.dart';
+import '../screens/splash/splash_screen.dart';
+
+class MyApp extends StatefulWidget {
+  @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  ThemeMode _themeMode = ThemeMode.dark;
+
+  @override
+  void initState() {
+    super.initState();
+    _loadThemePreference();
+  }
+
+  Future<void> _loadThemePreference() async {
+    final prefs = await SharedPreferences.getInstance();
+    final isDark = prefs.getBool('isDarkMode') ?? true;
+    setState(() {
+      _themeMode = isDark ? ThemeMode.dark : ThemeMode.light;
+    });
+  }
+
+  void _toggleTheme() async {
+    final newMode = _themeMode == ThemeMode.dark ? ThemeMode.light : ThemeMode.dark;
+    setState(() {
+      _themeMode = newMode;
+    });
+    final prefs = await SharedPreferences.getInstance();
+    await prefs.setBool('isDarkMode', newMode == ThemeMode.dark);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return ThemeProvider(
+      themeMode: _themeMode,
+      onThemeToggle: _toggleTheme,
+      child: MaterialApp(
+        title: 'InternTrack',
+        debugShowCheckedModeBanner: false,
+        theme: AppTheme.lightTheme,
+        darkTheme: AppTheme.darkTheme,
+        themeMode: _themeMode,
+        home: SplashScreen(),
+        routes: AppRoutes.routes,
+      ),
+    );
+  }
+}
