+++ lib/app/app_routes.dart
@@ -1,0 +1,17 @@
+// lib/app/app_routes.dart
+import 'package:flutter/material.dart';
+import '../screens/onboarding/onboarding_screen.dart';
+import '../screens/auth/auth_screen.dart';
+import '../screens/dashboard/student_dashboard.dart';
+
+class AppRoutes {
+  static const String onboarding = '/onboarding';
+  static const String auth = '/auth';
+  static const String dashboard = '/dashboard';
+
+  static Map<String, WidgetBuilder> routes = {
+    onboarding: (context) => OnboardingScreen(),
+    auth: (context) => AuthScreen(),
+    dashboard: (context) => StudentDashboard(),
+  };
+}
