+++ lib/core/constants/colors.dart
@@ -1,0 +1,16 @@
+// lib/core/constants/colors.dart
+import 'package:flutter/material.dart';
+
+class AppColors {
+  // Pure Black & White with Purple accent
+  static const Color pureBlack = Color(0xFF000000);
+  static const Color pureWhite = Color(0xFFFFFFFF);
+  static const Color darkGray = Color(0xFF0A0A0A);
+  static const Color lightGray = Color(0xFFF8F8F8);
+  static const Color mediumGray = Color(0xFF808080);
+  
+  // Purple accent - sophisticated and minimal
+  static const Color purplePrimary = Color(0xFF8B5CF6);
+  static const Color purpleLight = Color(0xFFA78BFA);
+  static const Color purpleDark = Color(0xFF7C3AED);
+}
