+++ lib/core/constants/strings.dart
@@ -1,0 +1,9 @@
+class AppStrings {
+  // App Info
+  static const String appName = 'InternTrack';
+  static const String appTagline = 'Your career journey, organized';
+  
+  // Keys
+  static const String onboardingKey = 'hasSeenOnboarding';
+  static const String themeKey = 'isDarkMode';
+}
