+++ lib/main.dart
@@ -1,13 +1,11 @@
+// lib/main.dart
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'firebase_options.dart';
-import 'core/theme/app_theme.dart';
-import 'core/providers/theme_provider.dart';
-import 'screens/splash/splash_screen.dart';
+import 'app/app.dart';
+import 'screens/dashboard/student_dashboard.dart';
 import 'screens/auth/auth_screen.dart';
-import 'screens/home_screen.dart';
-import 'package:shared_preferences/shared_preferences.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -15,54 +13,6 @@ void main() async {
     options: DefaultFirebaseOptions.currentPlatform,
   );
   runApp(MyApp());
-}
-
-class MyApp extends StatefulWidget {
-  @override
-  State<MyApp> createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  ThemeMode _themeMode = ThemeMode.dark;
-
-  @override
-  void initState() {
-    super.initState();
-    _loadThemePreference();
-  }
-
-  Future<void> _loadThemePreference() async {
-    final prefs = await SharedPreferences.getInstance();
-    final isDark = prefs.getBool('isDarkMode') ?? true;
-    setState(() {
-      _themeMode = isDark ? ThemeMode.dark : ThemeMode.light;
-    });
-  }
-
-  void _toggleTheme() async {
-    final newMode = _themeMode == ThemeMode.dark ? ThemeMode.light : ThemeMode.dark;
-    setState(() {
-      _themeMode = newMode;
-    });
-    final prefs = await SharedPreferences.getInstance();
-    await prefs.setBool('isDarkMode', newMode == ThemeMode.dark);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return ThemeProvider(
-      themeMode: _themeMode,
-      onThemeToggle: _toggleTheme,
-      child: MaterialApp(
-        title: 'InternTrack',
-        debugShowCheckedModeBanner: false,
-        theme: AppTheme.lightTheme,
-        darkTheme: AppTheme.darkTheme,
-        themeMode: _themeMode,
-        home: SplashScreen(),
-      ),
-    );
-  }
 }
 
 class AuthWrapper extends StatelessWidget {
@@ -84,7 +34,7 @@ class AuthWrapper extends StatelessWidget {
         }
         
         if (snapshot.hasData) {
-          return HomeScreen();
+          return StudentDashboard();
         }
         
         return AuthScreen();
