+++ lib/providers/theme_provider.dart
@@ -1,3 +1,4 @@
+// lib/providers/theme_provider.dart
 import 'package:flutter/material.dart';
 
 class ThemeProvider extends InheritedWidget {
