+++ lib/screens/dashboard/widgets/application_card.dart
@@ -1,0 +1,129 @@
+// lib/features/dashboard/widgets/application_card.dart
+import 'package:flutter/material.dart';
+import '../../../core/widgets/glass_container.dart';
+import '../../../core/constants/app_constants.dart';
+import '../../../core/constants/colors.dart';
+
+class ApplicationCard extends StatelessWidget {
+  final bool isDark;
+  final String company;
+  final String position;
+  final String status;
+  final String date;
+
+  const ApplicationCard({
+    Key? key,
+    required this.isDark,
+    required this.company,
+    required this.position,
+    required this.status,
+    required this.date,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return GlassContainer(
+      isDark: isDark,
+      padding: EdgeInsets.all(AppConstants.spaceL),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Flexible(
+                child: Row(
+                  children: [
+                    Container(
+                      width: 48,
+                      height: 48,
+                      decoration: BoxDecoration(
+                        gradient: LinearGradient(
+                          colors: [AppColors.purplePrimary, AppColors.purpleLight],
+                        ),
+                        borderRadius: BorderRadius.circular(AppConstants.radiusSmall),
+                      ),
+                      child: Icon(
+                        Icons.work_outline_rounded,
+                        color: Colors.white,
+                        size: 24,
+                      ),
+                    ),
+                    SizedBox(width: AppConstants.spaceM),
+                    Flexible(
+                      child: Column(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          Text(
+                            company,
+                            style: TextStyle(
+                              fontSize: 18,
+                              fontWeight: FontWeight.w700,
+                              color: isDark ? AppColors.pureWhite : AppColors.pureBlack,
+                            ),
+                            overflow: TextOverflow.ellipsis,
+                          ),
+                          SizedBox(height: 2),
+                          Text(
+                            position,
+                            style: TextStyle(
+                              fontSize: 14,
+                              color: AppColors.mediumGray,
+                            ),
+                            overflow: TextOverflow.ellipsis,
+                            maxLines: 1,
+                          ),
+                        ],
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+              SizedBox(width: 8),
+              Container(
+                padding: EdgeInsets.symmetric(
+                  horizontal: 12,
+                  vertical: 6,
+                ),
+                decoration: BoxDecoration(
+                  color: AppColors.purplePrimary.withOpacity(0.2),
+                  borderRadius: BorderRadius.circular(10),
+                  border: Border.all(
+                    color: AppColors.purplePrimary.withOpacity(0.5),
+                    width: 1.5,
+                  ),
+                ),
+                child: Text(
+                  status,
+                  style: TextStyle(
+                    fontSize: 11,
+                    fontWeight: FontWeight.w700,
+                    color: AppColors.purplePrimary,
+                  ),
+                ),
+              ),
+            ],
+          ),
+          SizedBox(height: AppConstants.spaceM),
+          Row(
+            children: [
+              Icon(
+                Icons.schedule_rounded,
+                size: 16,
+                color: AppColors.mediumGray,
+              ),
+              SizedBox(width: 6),
+              Text(
+                date,
+                style: TextStyle(
+                  fontSize: 13,
+                  color: AppColors.mediumGray,
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
